+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Created `Runner` and `Scraper` abstractions to retrieve and parse fox sports event data.
+- Created `NBAEventData` JSON response payload for fox sports NBA event data.
+- Created `NBABoxScoreStats` data model
+- Created `NBABoxScoreScraper` interface to scrape fox sports NBA box score data
+
+### Changed
+- Renamed `selectionId` to `segmentID`, `Id` to `ID`, and `SelectionId` to `SegmentID`
+- Documentation updates
 
 ## [0.4.0] - 2025-04-07
 ### Added